requirements.txt
pyspark==3.5.0
python-dotenv
supabase
pandas
